refactor: integrated corpus creation pipeline (#2791)
* pull wmg constants into separate file

* pull out extract functions for integrated corpus

* pull out transform functions for integrated corpus

* extract validation code for integrated corpus

* extract load code for integrated corpus

* refactor integrated corpus creation pipeline

* update tests

* use func timer
