+++ backend/wmg/data/cube_pipeline.py
@@ -1,17 +1,12 @@
-import gc
 import logging
-import os
 import pathlib
 import sys
 import time
-from typing import List
 import tiledb
 
-from backend.wmg.data import extract
+from backend.corpus_asset_pipelines import integrated_corpus
 from backend.wmg.data.load_cube import upload_artifacts_to_s3, make_snapshot_active
-from backend.wmg.data.load_corpus import load_h5ad
-from backend.wmg.data.schemas.corpus_schema import create_tdb, INTEGRATED_ARRAY_NAME
-from backend.wmg.data.tiledb import create_ctx
+from backend.wmg.data.schemas.corpus_schema import create_tdb
 from backend.wmg.data.transform import (
     generate_primary_filter_dimensions,
     get_cell_types_by_tissue,
@@ -22,30 +17,6 @@ from backend.wmg.data.wmg_cube import create_cubes
 
 logger = logging.getLogger(__name__)
 logging.basicConfig(level=logging.INFO)
-
-
-def load(dataset_directory: List, corpus_path: str, validate: bool = False):
-    """
-    Given the path to a directory containing one or more h5ad files and a group name, call the h5ad loading function
-    on all files, loading/concatenating the datasets together under the group name
-    """
-    with tiledb.scope_ctx(create_ctx()):
-        dataset_count = len(os.listdir(dataset_directory))
-        i = 0
-        for dataset in os.listdir(dataset_directory):
-            i += 1
-            logger.info(f"Processing dataset {i} of {dataset_count}")
-            h5ad_file_path = f"{dataset_directory}/{dataset}/local.h5ad"
-            load_h5ad(
-                h5ad_file_path, corpus_path, validate
-            )  # TODO Can this be parallelized? need to be careful handling global indexes but tiledb has a lock I think
-            gc.collect()
-
-        logger.info("all loaded, now consolidating.")
-
-        for arr_name in [f"{corpus_path}/{name}" for name in ["obs", "var", INTEGRATED_ARRAY_NAME]]:
-            tiledb.consolidate(arr_name)
-            tiledb.vacuum(arr_name)
 
 
 def load_data_and_create_cube(
@@ -71,12 +42,7 @@ def load_data_and_create_cube(
     if not tiledb.VFS().is_dir(corpus_path):
         create_tdb(snapshot_path, corpus_name)
 
-    if extract_data:
-        s3_uris = extract.get_dataset_s3_uris()
-        extract.copy_datasets_to_instance(s3_uris, path_to_h5ad_datasets)
-        logger.info("Copied datasets to instance")
-
-    load(path_to_h5ad_datasets, corpus_path, True)
+    integrated_corpus.run(path_to_h5ad_datasets, corpus_path, extract_data)
     logger.info("Loaded datasets into corpus")
     create_cubes(corpus_path)
     logger.info("Built expression summary cube")
